docs: reorganize /docs IA around 5 purpose-driven sections
Restructure the docs tree so each top-level section answers a distinct question
and sub-items are evenly balanced. The previous "How it Works" section mixed
two concerns (page-type deep dives + cross-cutting behaviours) and the landing
cards didn't match the sidebar sections.

New structure:

- Getting Started    ← "I'm new"
- Page Types         ← WHAT kinds of content exist (9 types as sub-pages)
- Features           ← HOW you use the content (pages, drives, AI, sharing,
                       search, accounts)
- Integrations       ← connecting external tools (MCP)
- Security & Trust   ← procurement / compliance surface

Moves (via git mv, history preserved):
- /docs/how-it-works/<type>/        -> /docs/page-types/<type>/  (9 pages)
- /docs/how-it-works/{pages,drives,sharing,search,accounts}/
                                    -> /docs/features/...        (5 pages)
- /docs/how-it-works/ai/            -> split into two pages:
    /docs/page-types/ai-chat/       (the page type)
    /docs/features/ai/              (cross-cutting AI behaviour)
- /docs/mcp/                        -> /docs/integrations/mcp/
- /docs/mcp/desktop/                -> /docs/integrations/mcp/desktop/

Deletes:
- /docs/ai/ (old "AI System" dev section; redundant with Features -> AI)
- /docs/how-it-works/ (empty after moves)

Other:
- Add /docs/features/page.tsx and /docs/integrations/page.tsx section landings.
- Rewrite /docs/page-types/page.tsx catalog links to point at the new URLs.
- Rewrite docs-nav.ts with the 5-section structure.
- Rewrite /docs/page.tsx landing: Start Here hero + 2x2 grid of the four
  reference sections.
- Update sitemap.ts for all new paths.
- Add 301 redirects /docs/mcp -> /docs/integrations/mcp in next.config.ts so
  external inbound links survive.
- Sweep every cross-link across the marketing app to the new URLs.
