+++ apps/marketing/next.config.ts
@@ -6,6 +6,12 @@ const nextConfig: NextConfig = {
   images: {
     path: '/_marketing/_next/image',
   },
+  async redirects() {
+    return [
+      { source: '/docs/mcp', destination: '/docs/integrations/mcp', permanent: true },
+      { source: '/docs/mcp/desktop', destination: '/docs/integrations/mcp/desktop', permanent: true },
+    ];
+  },
 };
 
 export default nextConfig;
