--- apps/marketing/src/app/docs/ai/page.tsx
@@ -1,100 +1,0 @@
-import { DocsMarkdown } from "@/components/DocsContent";
-import { createMetadata } from "@/lib/metadata";
-
-export const metadata = createMetadata({
-  title: "AI System",
-  description: "PageSpace AI architecture: AI conversations as pages, 12-provider support, 37 workspace tools, and agent collaboration.",
-  path: "/docs/ai",
-  keywords: ["AI", "LLM", "providers", "tool calling", "agents"],
-});
-
-const content = `
-# AI System
-
-AI conversations in PageSpace are pages, not a separate chatbot. They live inside drives and folders, inherit permissions, show up in search, and use the same tools that operate on the rest of your workspace.
-
-## AI conversations are pages
-
-Each AI chat is an \`AI_CHAT\` page type. It sits in the tree next to documents, folders, and channels:
-
-\`\`\`
-📁 Project Alpha/
-├── 📄 Requirements.md
-├── 📁 Research/
-│   ├── 🤖 Research Assistant    ← AI_CHAT page
-│   └── 📄 Market Analysis.md
-└── 🤖 Project Planning AI       ← AI_CHAT page
-\`\`\`
-
-Consequences:
-
-- AI inherits drive and page permissions — it can only see content the calling user can see.
-- AI conversations are searchable and mentionable like any other page.
-- Different parts of a workspace can host different agents with different configs.
-
-## Message persistence
-
-Every message — user or assistant — is persisted. Messages carry their author, role, text, and full tool-call history; an agent that ran a tool hours ago can still explain what it did. Persistence also enables multi-user conversations (several people sharing one agent) and real-time sync across connected clients.
-
-## Context inheritance
-
-Every tool call receives an execution context with the caller's user, timezone, active provider and model, and the page and drive it is running inside. Agents use this context to scope search, resolve "here" to the right page, and pick the correct drive for new pages.
-
-## Providers and models
-
-PageSpace routes AI through the Vercel AI SDK across 12 providers:
-
-| Provider | Default model | Who supplies the key |
-|---|---|---|
-| PageSpace | \`glm-4.7\` (Standard), \`glm-5\` (Pro) | Built-in (GLM backend) |
-| OpenRouter (Paid) | user-selected | User |
-| OpenRouter (Free) | user-selected | User (no cost) |
-| Google AI | \`gemini-2.5-flash\` | User |
-| OpenAI | user-selected | User |
-| Anthropic | user-selected | User |
-| xAI (Grok) | user-selected | User |
-| GLM (Coder Plan) | user-selected | User |
-| MiniMax | user-selected | User |
-| Ollama | discovered from local instance | N/A (local server) |
-| LM Studio | discovered from local instance | N/A (local server) |
-| Azure OpenAI | user's deployment name | User |
-
-The full list of available models is visible in **Settings > AI** and updates as providers publish new ones.
-
-## Tools
-
-AI has access to 37 workspace tools. They cover page reads and writes, drive management, search, calendar, channels, tasks, agent discovery and consultation, activity queries, and web search. See [Tool Calling](/docs/ai/tool-calling) for the full reference.
-
-Tools run with the calling user's permissions. An AI cannot read, write, or delete anything the user cannot.
-
-## Read-only and web-search toggles
-
-Each AI_CHAT page exposes two runtime toggles that filter the tool set:
-
-- **Read-only mode** — excludes every write tool (create, edit, delete, send). The agent can still read, search, and plan.
-- **Web search enabled** — includes \`web_search\`; off by default.
-
-A page can also pin a specific subset of tools in its configuration.
-
-## Agent-to-agent consultation
-
-An agent can call another agent via \`ask_agent\`, passing a question and optional context. The target agent loads its own conversation history, runs with its own system prompt and tools, persists the exchange, and returns a response. Chain depth is bounded to prevent runaway agent-calling-agent loops.
-
-## Real-time collaboration
-
-AI messages are broadcast over Socket.IO to every user viewing the AI_CHAT page. Tool calls, tool results, and streaming response deltas all reach connected clients. Multi-user chat with a single agent works out of the box.
-
-## Model capability detection
-
-Before running a tool-capable workflow, PageSpace checks whether the selected model supports tool calling and vision. If tools aren't supported, the stream falls back to text-only and the UI suggests tool-capable alternatives from the same provider. OpenRouter capability is queried at runtime; other providers use pattern matching on the model id.
-
-## Learn more
-
-- **[Providers & Models](/docs/ai/providers)** — per-provider setup, model ids, and API-key storage.
-- **[Tool Calling](/docs/ai/tool-calling)** — the 37-tool catalog, execution context, retries, and custom tool sets.
-- **[Agents](/docs/ai/agents)** — configuring AI_CHAT pages, system prompts, and agent-to-agent flows.
-`;
-
-export default function AIPage() {
-  return <DocsMarkdown content={content} />;
-}
