+++ apps/marketing/src/app/docs/docs-nav.ts
@@ -1,9 +1,6 @@
 import {
   Book,
   Sparkles,
-  Cpu,
-  Bot,
-  Wrench,
   Server,
   Monitor,
   Shield,
@@ -23,6 +20,9 @@ import {
   Palette,
   Table,
   Code,
+  Bot,
+  Layers,
+  Plug,
 } from "lucide-react";
 
 export interface NavItem {
@@ -45,53 +45,51 @@ export const docsNav: NavSection[] = [
       { title: "Overview", href: "/docs" },
       { title: "Quick Start", href: "/docs/getting-started", icon: Book },
       { title: "Core Concepts", href: "/docs/core-concepts", icon: Blocks },
-      { title: "Page Types", href: "/docs/page-types", icon: LayoutGrid },
     ],
   },
   {
-    title: "How it Works",
+    title: "Page Types",
+    icon: LayoutGrid,
+    items: [
+      { title: "Overview", href: "/docs/page-types", icon: LayoutGrid },
+      { title: "Document", href: "/docs/page-types/document", icon: FileText },
+      { title: "Folder", href: "/docs/page-types/folder", icon: Folder },
+      { title: "AI Chat", href: "/docs/page-types/ai-chat", icon: Bot },
+      { title: "Channel", href: "/docs/page-types/channel", icon: MessageSquare },
+      { title: "Sheet", href: "/docs/page-types/sheet", icon: Table },
+      { title: "Canvas", href: "/docs/page-types/canvas", icon: Palette },
+      { title: "Code", href: "/docs/page-types/code", icon: Code },
+      { title: "Task List", href: "/docs/page-types/task-list", icon: ListChecks },
+      { title: "File", href: "/docs/page-types/file", icon: Upload },
+    ],
+  },
+  {
+    title: "Features",
     icon: Lightbulb,
     items: [
-      { title: "Overview", href: "/docs/how-it-works", icon: Lightbulb },
-      { title: "Pages", href: "/docs/how-it-works/pages", icon: FileText },
-      { title: "Documents", href: "/docs/how-it-works/documents", icon: FileText },
-      { title: "Folders", href: "/docs/how-it-works/folders", icon: Folder },
-      { title: "AI in your Workspace", href: "/docs/how-it-works/ai", icon: Sparkles },
-      { title: "Channels", href: "/docs/how-it-works/channels", icon: MessageSquare },
-      { title: "Task Lists", href: "/docs/how-it-works/task-lists", icon: ListChecks },
-      { title: "Sheets", href: "/docs/how-it-works/sheets", icon: Table },
-      { title: "Canvas", href: "/docs/how-it-works/canvas", icon: Palette },
-      { title: "Code", href: "/docs/how-it-works/code", icon: Code },
-      { title: "Files & Uploads", href: "/docs/how-it-works/files", icon: Upload },
-      { title: "Drives & Workspaces", href: "/docs/how-it-works/drives", icon: HardDrive },
-      { title: "Search", href: "/docs/how-it-works/search", icon: Search },
-      { title: "Sharing & Permissions", href: "/docs/how-it-works/sharing", icon: Users },
-      { title: "Accounts & Sign In", href: "/docs/how-it-works/accounts", icon: KeyRound },
+      { title: "Overview", href: "/docs/features", icon: Lightbulb },
+      { title: "Pages", href: "/docs/features/pages", icon: Layers },
+      { title: "Drives & Workspaces", href: "/docs/features/drives", icon: HardDrive },
+      { title: "AI in your Workspace", href: "/docs/features/ai", icon: Sparkles },
+      { title: "Sharing & Permissions", href: "/docs/features/sharing", icon: Users },
+      { title: "Search", href: "/docs/features/search", icon: Search },
+      { title: "Accounts & Sign In", href: "/docs/features/accounts", icon: KeyRound },
     ],
   },
   {
-    title: "AI System",
-    icon: Sparkles,
+    title: "Integrations",
+    icon: Plug,
     items: [
-      { title: "AI Overview", href: "/docs/ai", icon: Sparkles },
-      { title: "Providers & Models", href: "/docs/ai/providers", icon: Cpu },
-      { title: "Tool Calling", href: "/docs/ai/tool-calling", icon: Wrench },
-      { title: "Agents", href: "/docs/ai/agents", icon: Bot },
+      { title: "Overview", href: "/docs/integrations", icon: Plug },
+      { title: "MCP", href: "/docs/integrations/mcp", icon: Server },
+      { title: "Desktop MCP", href: "/docs/integrations/mcp/desktop", icon: Monitor },
     ],
   },
   {
-    title: "MCP Integration",
-    icon: Server,
-    items: [
-      { title: "MCP Overview", href: "/docs/mcp", icon: Server },
-      { title: "Desktop MCP", href: "/docs/mcp/desktop", icon: Monitor },
-    ],
-  },
-  {
-    title: "Security & Auth",
+    title: "Security & Trust",
     icon: Shield,
     items: [
-      { title: "Security Overview", href: "/docs/security", icon: Shield },
+      { title: "Overview", href: "/docs/security", icon: Shield },
       { title: "Authentication", href: "/docs/security/authentication", icon: KeyRound },
       { title: "Permissions", href: "/docs/security/permissions", icon: Users },
       { title: "Zero-Trust", href: "/docs/security/zero-trust", icon: Eye },
