+++ apps/marketing/src/app/docs/getting-started/page.tsx
@@ -115,7 +115,7 @@ If you use Claude Desktop, Claude Code, Cursor, or other MCP-compatible tools, y
 }
 \`\`\`
 
-See the [MCP Integration guide](/docs/mcp) for full details.
+See the [MCP Integration guide](/docs/integrations/mcp) for full details.
 
 ## 7. Invite Your Team
 
@@ -131,9 +131,9 @@ Members can collaborate in real-time on documents, channels, and AI conversation
 ## Next Steps
 
 - **[Core Concepts](/docs/core-concepts)** — pages as primitives, context inheritance, the tree model
-- **[How it Works](/docs/how-it-works)** — feature-by-feature plain-language reference
-- **[Page Types](/docs/page-types)** — the 9 built-in page types at a glance
-- **[MCP Integration](/docs/mcp)** — connect Claude Desktop, Cursor, or your own MCP client
+- **[Page Types](/docs/page-types)** — the 9 built-in page types and what each one is for
+- **[Features](/docs/features)** — plain-language reference for the behaviours every page shares
+- **[MCP Integration](/docs/integrations/mcp)** — connect Claude Desktop, Cursor, or your own MCP client
 `;
 
 export default function GettingStartedPage() {
