+++ apps/marketing/src/app/sitemap.ts
@@ -63,32 +63,30 @@ export default function sitemap(): MetadataRoute.Sitemap {
     "/docs",
     "/docs/getting-started",
     "/docs/core-concepts",
+    // Page Types
     "/docs/page-types",
-    // How it Works
-    "/docs/how-it-works",
-    "/docs/how-it-works/pages",
-    "/docs/how-it-works/documents",
-    "/docs/how-it-works/folders",
-    "/docs/how-it-works/ai",
-    "/docs/how-it-works/channels",
-    "/docs/how-it-works/task-lists",
-    "/docs/how-it-works/sheets",
-    "/docs/how-it-works/canvas",
-    "/docs/how-it-works/code",
-    "/docs/how-it-works/files",
-    "/docs/how-it-works/drives",
-    "/docs/how-it-works/search",
-    "/docs/how-it-works/sharing",
-    "/docs/how-it-works/accounts",
-    // AI
-    "/docs/ai",
-    "/docs/ai/providers",
-    "/docs/ai/tool-calling",
-    "/docs/ai/agents",
-    // MCP
-    "/docs/mcp",
-    "/docs/mcp/desktop",
-    // Security
+    "/docs/page-types/document",
+    "/docs/page-types/folder",
+    "/docs/page-types/ai-chat",
+    "/docs/page-types/channel",
+    "/docs/page-types/sheet",
+    "/docs/page-types/canvas",
+    "/docs/page-types/code",
+    "/docs/page-types/task-list",
+    "/docs/page-types/file",
+    // Features
+    "/docs/features",
+    "/docs/features/pages",
+    "/docs/features/drives",
+    "/docs/features/ai",
+    "/docs/features/sharing",
+    "/docs/features/search",
+    "/docs/features/accounts",
+    // Integrations
+    "/docs/integrations",
+    "/docs/integrations/mcp",
+    "/docs/integrations/mcp/desktop",
+    // Security & Trust
     "/docs/security",
     "/docs/security/authentication",
     "/docs/security/permissions",
