+++ apps/marketing/src/components/SiteFooter.tsx
@@ -14,7 +14,7 @@ const resourceLinks = [
   { label: "Documentation", href: "/docs" },
   { label: "Getting Started", href: "/docs/getting-started" },
   { label: "API Reference", href: "/docs/api" },
-  { label: "MCP Integration", href: "/docs/mcp" },
+  { label: "MCP Integration", href: "/docs/integrations/mcp" },
 ];
 
 const companyLinks = [
